requirements.txt
gensim
umap-learn
plotly
